+++ BambooTracker/bamboo_tracker.cpp
@@ -685,7 +685,7 @@ void BambooTracker::jamKeyOn(JamKey key)
 	auto attrib = songStyle_.trackAttribs[static_cast<size_t>(curTrackNum_)];
 	if (attrib.source == SoundSource::Drum) {
 		opnaCtrl_->setKeyOnFlagDrum(attrib.channelInSource);
-		opnaCtrl_->updateKeyOnOffStatusDrum();
+		opnaCtrl_->updateRegisterStates();
 	}
 	else {
 		int keyNum = octaveAndNoteToNoteNumber(octave_, JamManager::jamKeyToNote(key));
@@ -749,7 +749,7 @@ void BambooTracker::jamKeyOn(int keyNum)
 	auto attrib = songStyle_.trackAttribs[static_cast<size_t>(curTrackNum_)];
 	if (attrib.source == SoundSource::Drum) {
 		opnaCtrl_->setKeyOnFlagDrum(attrib.channelInSource);
-		opnaCtrl_->updateKeyOnOffStatusDrum();
+		opnaCtrl_->updateRegisterStates();
 	}
 	else {
 		std::vector<JamKeyData>&& list = jamMan_->keyOn(JamKey::MIDI_KEY, attrib.channelInSource, attrib.source, keyNum);
@@ -811,7 +811,7 @@ void BambooTracker::jamKeyOff(JamKey key)
 	auto attrib = songStyle_.trackAttribs[static_cast<size_t>(curTrackNum_)];
 	if (attrib.source == SoundSource::Drum) {
 		opnaCtrl_->setKeyOffFlagDrum(attrib.channelInSource);
-		opnaCtrl_->updateKeyOnOffStatusDrum();
+		opnaCtrl_->updateRegisterStates();
 	}
 	else {
 		int keyNum = octaveAndNoteToNoteNumber(octave_, JamManager::jamKeyToNote(key));
@@ -845,7 +845,7 @@ void BambooTracker::jamKeyOff(int keyNum)
 	auto attrib = songStyle_.trackAttribs[static_cast<size_t>(curTrackNum_)];
 	if (attrib.source == SoundSource::Drum) {
 		opnaCtrl_->setKeyOffFlagDrum(attrib.channelInSource);
-		opnaCtrl_->updateKeyOnOffStatusDrum();
+		opnaCtrl_->updateRegisterStates();
 	}
 	else {
 		JamKeyData&& data = jamMan_->keyOff(JamKey::MIDI_KEY, keyNum);